Remove reliance on System.Drawing.Color type
I had added a hashtable of RGB values defined using System.Drawing.Color
to help convert temperatures in to colours for smartbulbs like Hue. It
really didn't make sense for it to be in this class so I've removed it
and put it in PoSHue instead. This has the added benefit of removing the
need for Add-Type -AssemblyName System.Drawing.Color before every
script!