SaveActivePlugScheduleToFile and SetActivePlugScheduleFromFile
Encrypted the password in the object now as having it visible in clear
text wasn't great. I'd like to improve this further but I don't want it
to be a barrier to adoption of the module if users have to use
ConvertTo-SecureString -String "password" -AsPlainText -Force just to
set their password. Would be a breaking change for existing scripts.
As promised, have implemented SaveActivePlugScheduleToFile and
SetActivePlugScheduleFromFile and bumped to 1.4.1.